moved notification testing into /notifications
